requirements.txt
pycryptodome
qrcode
pypng
zbar
pillow